backend/requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-mail==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
openai==1.35.7
orjson==3.10.5
passlib==1.7.4
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
